Remove amount-based confirmation scaling (#3044)
We previously scaled the number of confirmations based on the channel
capacity: however, this doesn't really work with splicing, as shown by
the following attack scenario:

- a malicious node opens a very large channel
- we thus wait for more confirmations to protect against reorgs
- they then send lightning payments until all of the balance is on our
  side of the channel
- we splice-out to shrink the channel to a minimal size
- if we thus us a smaller number of confirmations (because the channel
  is now small), the malicious node can perform this small reorg and
  publish one of their revoked commitment

There are more involved variations of this attack, which show that the
real amount at stake is potentially much more than a single channel's
current capacity. Instead of trying to find a complex solution that
would likely have subtle edge cases, it's a lot simpler to always use
a static number of confirmations that is large enough to protect against
reorgs entirely.

+++ eclair-core/src/main/scala/fr/acinq/eclair/channel/fsm/CommonFundingHandlers.scala
@@ -55,7 +55,7 @@ trait CommonFundingHandlers extends CommonHandlers {
   /**
    * @param delay_opt optional delay to reduce herd effect at startup.
    */
-  def watchFundingConfirmed(fundingTxId: TxId, minDepth_opt: Option[Long], delay_opt: Option[FiniteDuration]): Unit = {
+  def watchFundingConfirmed(fundingTxId: TxId, minDepth_opt: Option[Int], delay_opt: Option[FiniteDuration]): Unit = {
     val watch = minDepth_opt match {
       case Some(fundingMinDepth) => WatchFundingConfirmed(self, fundingTxId, fundingMinDepth)
       // When using 0-conf, we make sure that the transaction was successfully published, otherwise there is a risk
